Move classes for llk parameters to correct file (llk_params.py)
